requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
WTForms==3.0.1
Werkzeug==2.3.7
SQLAlchemy==2.0.21
openai==1.3.5
python-dotenv==1.0.0
bcrypt==4.0.1
pandas==2.1.1
numpy==1.25.2
plotly==5.17.0
gunicorn==21.2.0
